Se utilizan los metodos ya creados para realizar el reducer
