--- src/store/index.js
@@ -1,9 +1,0 @@
-import { configureStore } from "@reduxjs/toolkit";
-// reducers
-import slides from './slices/slides';
-
-export default configureStore({
-    reducer: {
-        slides
-    }
-});
